Incorporate comments

+++ src/cudamatrix/cu-allocator.h
@@ -338,8 +338,8 @@ class CuMemoryAllocator {
 
   // Keep track of the memory usage from the cache to track the maximum memory used by
   //   the application
-  int max_allocated_memory_;
-  int allocated_memory_;
+  size_t max_allocated_memory_;
+  size_t allocated_memory_;
 };
